ChatBot
This is chatbot made Purely via Python . This chatbot is designed for a user to get their all queries related to National Scholarship portal by Indian government Chatbots can be a powerful tool for improving user experiences. They are commonly used for customer support, website help, and more. They are very helpful in saving time as the answers to the users will be directed in seconds .

For chatbots to work well, you need structured content. There are three main elements to chatbot interactions: context, intent, entity.

Context - the reason for the interaction

Intent - the purpose or goal of the interaction

Entity - delivering the content that is required as a result of the interaction (this is where structured content comes into play) Any entity can be a granular component of information, or if it's a larger piece of content, it can be semantically tagged. The chatbot can then interface with it and pull out a smaller piece of information it requires.” Proposed Solution For making the process fast the FAQs can be answered using the latest AI technology by python where the use of API "chatterbot" can automatically generate according to the questions. Also for every question asked it can be stored and the list of questions can be increased to provide a better mapping of questions and answers with the use of the DATABASE MANAGEMENT SYSTEM. Along with the Backend of Python, we can also give it a better GUI according to our needs and the topic of FAQs using python flash CSS div classes to give our project a better overview.